Added VS Code Support on Windows

+++ install.py
@@ -11,8 +11,10 @@ MACOS = 'macos'
 UNIX = 'unix'
 FILES_DIR = 'files'
 CONFIG_DIR = 'config'
+VSCODE_DIR = 'vscode'
 HOME = str(Path.home())
 CWD = os.getcwd()
+CODEWINDOWS = os.path.join(HOME, "AppData", "Roaming", "Code", "User")
 
 def main():
     with open('data.json') as f:
@@ -24,20 +26,29 @@ def main():
             platform_name = MACOS
         elif os_type.startswith('win32'):
             platform_name = WINDOWS
+            setupVSCode(CODEWINDOWS)
+            fs_type = WINDOWS
         elif os_type.startswith('linux'):
             platform_name = LINUX
-
-        if platform_name == WINDOWS:
-            fs_type = WINDOWS
 
         files = data['files']['shared'] + data['files'][platform_name]
         folders = data['folders']['shared'] + data['folders'][platform_name]
         config = data['config']['shared'] + data['config'][platform_name]
 
+
         if fs_type == UNIX:
             return unix(files, folders, config)
 
         return windows(files, folders, config)
+
+def setupVSCode(CODE):
+    repoDir = os.path.join(CWD, VSCODE_DIR)
+    for f in os.listdir(repoDir):
+        fullPath = os.path.join(CWD, VSCODE_DIR, f)
+        codePath = os.path.join(CODE, f)
+        if os.path.exists(fullPath) and not os.path.exists(codePath):
+            os.symlink(fullPath, codePath)
+ 
 
 def unix(files, folders, config):
     hiddenConfig = hidden(CONFIG_DIR, UNIX)
@@ -68,15 +79,14 @@ def windows(files, folders, config):
         fullPath = os.path.join(CWD, FILES_DIR, f)
         hf = hidden(f, UNIX)
         homePath = os.path.join(HOME, hf)
+        actualPath = None
         if f == 'Microsoft.PowerShell_profile.ps1':
             actualPath = os.path.join(HOME, 'Documents', 'WindowsPowerShell', f)
-            if os.path.isfile(fullPath) and not os.path.isfile(actualPath):
-                os.symlink(fullPath, actualPath)
         elif f == "wtprofile.json":
             actualPath = os.path.join(HOME, "AppData", "Local", "Packages", \
                     "Microsoft.WindowsTerminal_8wekyb3d8bbwe", "LocalState", \
                     "profiles.json")
-            if os.path.isfile(fullPath) and not os.path.isfile(actualPath):
+        if actualPath and os.path.isfile(fullPath) and not os.path.isfile(actualPath):
                 os.symlink(fullPath, actualPath)
         if os.path.isfile(fullPath) and not os.path.isfile(homePath):
             os.symlink(fullPath, homePath)
